Fix font inside block body
Prior to this patch, LaTeX ``double quotes'' were not working properly
inside the block body environment: they rendered as the literal
characters "``" and "''" rather than as smart quotes.

This patch removes the explicit setting of the font for the block body
environment, which seems to fix the issue with the double quotes. This
explicit setting was unnecessary anyways, because of the `\setsansfont`
command that sent the font for the document.