Switch to built-in Beamer caption template
Beamer has a built in template for numbered captions, so there is no
reason to be using a custom template.

Furthermore, the old code caused some issues with newer versions of
Beamer (tested with v3.55). Captions rendered like this:

    Figure 1:Example

However, the expected behavior, as well as the observed behavior with
older versions of Beamer (tested with v3.41), was:

    Figure 1: Example

This change seems to fix the behavior across versions; plus, it
simplifies the code.