Fix alert block background width
Newer versions of Beamer (tested with v3.55) introduced some change that
broke the rendering of alert blocks, which rendered properly with older
versions of Beamer (tested with v3.41).

Josh Holland <anowlcalledjosh@gmail.com> found that removing the vmode
option from the beamercolorbox environment fixes the problem.

It is unclear why the vmode option was used in the first place. The
Beamer manual says that the option causes TeX to be in vertical mode
when the box starts (where normally TeX would be in horizontal mode),
and that "only TeXperts need this option, so, if you use it, you will
probably know what you are doing anyway."

In this situation, there was no particular reason why the option was
being used, so this patch removes it. Simply removing the vmode option
affects the spacing between the header's horizontal line and the start
of the text, so this patch adjusts the spacing to compensate.

The change seems to fix the behavior across versions.